src/services/gamification.service.ts
import type { UserRole, ParticipationType, IUserGameProfile } from '../types/gamification';
import { CacheService } from './cache.service';
import { Injectable } from '@nestjs/common';
import { MonitoringService } from './monitoring.service';

@Injectable()
export class GamificationService {
  constructor(
    private readonly cacheService: CacheService,
    private readonly monitoringService: MonitoringService,
  ) {}

  private async createUserProfile(userId: string, role: UserRole): Promise<IUserGameProfile> {
    // 实现创建用户档案逻辑
    return {} as IUserGameProfile;
  }

  private async assignInitialChallenges(userId: string, role: UserRole): Promise<void> {
    // 实现分配初始挑战逻辑
  }

  private async setupAchievements(userId: string, role: UserRole): Promise<void> {
    // 实现设置成就逻辑
  }

  private async updateUserStats(userId: string, action: string, value: number): Promise<void> {
    // 实现更新用户统计逻辑
  }

  private async checkAchievements(userId: string, role: UserRole): Promise<void> {
    // 实现检查成就逻辑
  }

  private async updateChallenges(userId: string): Promise<void> {
    // 实现更新挑战逻辑
  }

  private async getUserProfile(userId: string): Promise<IUserGameProfile> {
    // 实现获取用户档案逻辑
    return {} as IUserGameProfile;
  }
}
